Extracted template and logic from task-form component
