+++ src/components/task-form/index.js
@@ -1,0 +1,8 @@
+import Component from "../component";
+import {template} from "./template";
+
+export default class TaskForm extends Component {
+  constructor(task) {
+    super(template(task));
+  }
+}
